requirements.txt
python-dotenv
schedule
gotify
cryptography
requests